delete_all(user_uid: user_uid) if where(user_uid: user_uid).any? in TwitterDB::Friendship

+++ app/models/twitter_db/followership.rb
@@ -9,7 +9,7 @@ module TwitterDB
       followerships = follower_uids.map.with_index { |follower_uid, i| [user_uid, follower_uid, i] }
 
       ActiveRecord::Base.transaction do
-        delete_all(user_uid: user_uid)
+        delete_all(user_uid: user_uid) if where(user_uid: user_uid).any?
         import(%i(user_uid follower_uid sequence), followerships, validate: false, timestamps: false)
       end
     end
