+++ app/models/twitter_db/friendship.rb
@@ -9,7 +9,7 @@ module TwitterDB
       friendships = friend_uids.map.with_index { |friend_uid, i| [user_uid, friend_uid, i] }
 
       ActiveRecord::Base.transaction do
-        delete_all(user_uid: user_uid)
+        delete_all(user_uid: user_uid) if where(user_uid: user_uid).any?
         import(%i(user_uid friend_uid sequence), friendships, validate: false, timestamps: false)
       end
     end
